requirements.txt
blinker==1.7.0
click==8.1.7
Flask==2.2.3
Flask-WTF==0.15.1
Flask-Migrate==3.0.1
SQLAlchemy==1.4.32
psycopg2==2.9.1
Pillow==9.0.0
Flask-Flash
gunicorn==21.2.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.2
packaging==23.2
python-dotenv==1.0.1
Werkzeug==2.2.3
zipp==3.13.0
email_validator==1.1.3
Flask-Mail==0.9.1
WTForms==3.0.1